requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
deep-translator==1.11.4
idna==3.6
lxml==4.9.3
packaging==23.2
Pillow==10.1.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
python-pptx==0.6.23
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
XlsxWriter==3.1.9
